+++ src/main/kotlin/com/github/gib/services/MyApplicationService.kt
@@ -1,6 +1,6 @@
-package com.github.jonatha1983.gib.services
+package com.github.gib.services
 
-import com.github.jonatha1983.gib.MyBundle
+import com.github.gib.MyBundle
 
 class MyApplicationService {